Move IS_DEV to constants
